# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/evidence:
    post:
      operationId: createPersistentEvidence
      tags: [cashnet]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/EvidenceInput" } } } }
      responses: { "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/PersistentEvidence" } } } } }
  /v1/evidence/{id}:
    get:
      operationId: getPersistentEvidence
      tags: [cashnet]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: Evidence, content: { application/json: { schema: { $ref: "#/components/schemas/PersistentEvidence" } } } } }
components:
  schemas:
...
    PersistentEvidence:
      type: object
      required: [id, subjectType, subjectId, evidenceType, sourceType, createdAt]
      properties:
        id: { type: string }
        caseId: { type: [string, "null"] }
        investigationId: { type: [string, "null"] }
        subjectType: { type: string }
        subjectId: { type: string }
        evidenceType: { type: string }
        sourceType: { type: string }
        provider: { type: [string, "null"] }
        sourceReference: { type: [string, "null"] }
        sourceUrl: { type: [string, "null"] }
        observedAt: { type: [string, "null"], format: date-time }
        collectedAt: { type: [string, "null"], format: date-time }
        method: { type: [string, "null"] }
        confidence: { type: [number, "null"], minimum: 0, maximum: 1 }
        rawReference: { type: [string, "null"] }
        contentHash: { type: [string, "null"] }
        description: { type: [string, "null"] }
        createdBy: { type: [string, "null"] }
        createdAt: { type: string, format: date-time }
    EvidenceInput:
      type: object
      required: [caseId, subjectType, subjectId, evidenceType, sourceType]
      properties:
        caseId: { type: string }
        investigationId: { type: [string, "null"] }
        subjectType: { type: string }
        subjectId: { type: string }
        evidenceType: { type: string, enum: [BLOCKCHAIN_FACT, TRANSACTION, ADDRESS_LABEL, ENTITY_MATCH, VASP_MATCH, GRAPH_RELATION, RISK_INDICATOR, DOCUMENT, OSINT, OTHER] }
        sourceType: { type: string, enum: [SYNTHETIC, API, RPC, DATASET, INFERENCE, OTHER, USER_PROVIDED] }
        provider: { type: [string, "null"] }
        sourceReference: { type: [string, "null"] }
        sourceUrl: { type: [string, "null"] }
        observedAt: { type: [string, "null"], format: date-time }
        collectedAt: { type: [string, "null"], format: date-time }
        method: { type: [string, "null"] }
        confidence: { type: [number, "null"], minimum: 0, maximum: 1 }
        rawReference: { type: [string, "null"] }
        contentHash: { type: [string, "null"] }
        description: { type: [string, "null"] }
...